More clean up and foxfordng
Former-commit-id: 4d2800004a29f061c030dcd6e9db305cdfac56c6 [formerly 764d32c3599103e91bd2213101e741de057b3cdd]
Former-commit-id: 7f887a7f3d5d22d1763a474ee7bcd088de4c9910 [formerly 4eff583e5821013119207c062558517afff96e2b]
Former-commit-id: e0f138119d6434354cde102cd6b76873ab4ad909
Former-commit-id: 1f5b81306dc8ed9c8d3718dce82ad6fd13c9d2f5

+++ Source/RawParser/Model/Parser/Decoder/DngDecoder.cs
@@ -74,7 +74,7 @@ namespace RawNet
             for (int k = data.Count - 1; k >= 0; k--)
             {
                 IFD i = data[k];
-                int comp = i.GetEntry(TagType.COMPRESSION).GetShort(0);
+                int comp = i.GetEntry(TagType.COMPRESSION).GetInt(0);
                 bool isSubsampled = false;
                 try
                 {
